Try overriding the parseErrorResponse method to munge the errors into the right format
